Fix compilation.
For some reason, CMake (or more specifically the Qt module) decided to append -std=gnu++11
which has overriden our current setting of -std=c++14 and breaking compilation.

The solution is to use the new CMAKE_CXX_STANDARD variable. As a bonus, this variables is ought
to be portable among compilers.